Refactoring of how time bar is displayed in each test and other minor changes
