+++ src/components/Header/Header.module.css
@@ -11,31 +11,6 @@
 .header nav {
 	position: absolute;
 	left: 20px;
-	display: flex;
-	align-items: center;
-	background-color: white;
-	font-size: 1.2rem;
-}
-
-.header nav ul {
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-}
-.header nav ul li {
-	width: fit-content;
-	white-space: nowrap;
-	list-style: none;
-	margin: 0 25px;
-	padding: 5px;
-}
-.header {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	position: relative;
-	height: 60px;
-	border-bottom: 1px solid #ccc;
 }
 
 .title {
@@ -44,33 +19,4 @@
 	transform: translateX(-50%);
 	font-size: 20px;
 	font-weight: bold;
-}
-
-.toggle_group {
-	border: 1px solid #ccc;
-	border-radius: 10px;
-
-	overflow: hidden;
-	background-color: white;
-	gap: 0; /* removes spacing between items */
-}
-
-.toggle_group_label {
-	margin-right: 20px;
-}
-
-.activeLink {
-	background-color: #5d92c0;
-	color: white;
-	padding: 10px 20px;
-	text-decoration: none;
-	transition: all 0.3s ease;
-}
-
-.inactiveLink {
-	background-color: white;
-	color: grey;
-	padding: 10px 20px;
-	text-decoration: none;
-	transition: all 0.3s ease;
 }
