+++ src/components/QuestionToggleSwitch/QuestionToggleSwitch.module.css
@@ -1,0 +1,35 @@
+.container {
+	display: flex;
+	align-items: center;
+	background-color: white;
+	font-size: 1.2rem;
+}
+
+.toggle_group {
+	border: 1px solid #ccc;
+	border-radius: 10px;
+	overflow: hidden;
+	background-color: white;
+	font-size: 1rem;
+	padding: 7px 0;
+}
+
+.toggle_group_label {
+	margin-right: 10px;
+}
+
+.activeLink {
+	background-color: #5d92c0;
+	color: white;
+	padding: 18px 20px;
+	text-decoration: none;
+	transition: all 0.3s ease;
+}
+
+.inactiveLink {
+	background-color: white;
+	color: grey;
+	padding: 18px 20px;
+	text-decoration: none;
+	transition: all 0.3s ease;
+}
