+++ src/components/SpeakingPart1/SpeakingPart1.module.css
@@ -1,0 +1,17 @@
+.choices {
+	font-size: 1.4rem;
+	margin-left: 25px;
+}
+
+.option {
+	margin: 8px;
+}
+
+.tags_container {
+	margin-top: 30px;
+}
+
+.instruction {
+	font-size: 1rem;
+	margin-left: 10px;
+}
